Split Pandas to Scalar, Grouped Map, Grouped Agg, Window
